refactor: domain 폴더 제거
