+++ panda-market/app/board/[id]/ArticleItem.tsx
