+++ panda-market/app/boards/BestArticleList.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Article, getData } from '@/app/boards/page';
-import BestArticleItem from '@/components/domain/best-articles/BestArticleItem';
+import BestArticleItem from '@/app/boards/BestArticleItem';
 import { useMediaQuery } from 'react-responsive';
 import { useEffect, useState } from 'react';
